Pass ruff 0.16: fix the lint findings its newer rules add.
- Prefix unpacked values the tests never read with an underscore (RUF059).
- Declare generic functions with type parameters (UP047, UP049) and drop the
  TypeVars nothing uses any more.
- Drop int() around math.ceil() and round(), which already return ints (RUF046).
- Resolve the retirement migration's path at import time, not inside an async
  test (ASYNC240).
- Sort two import blocks, drop an unused noqa, and apply 0.16's formatting to
  three files.

+++ apps/api/app/modules/math/solve/graph.py
@@ -168,7 +168,7 @@ def _adapt_sample_window(
     if needed <= n:
         return x_min, x_max, n
     if needed <= _MAX_GRAPH_SAMPLES:
-        return x_min, x_max, min(_MAX_GRAPH_SAMPLES, max(n, int(math.ceil(needed))))
+        return x_min, x_max, min(_MAX_GRAPH_SAMPLES, max(n, math.ceil(needed)))
     mid = 0.0 if x_min < 0 < x_max else (x_min + x_max) / 2.0
     half = (_OSCILLATION_PERIODS_IN_VIEW * period) / 2.0
     zoom_n = min(
